move a star struct for a better understanding
